-done unit test service (except for manga and chapter service) -add env support for database connection on migration app -add template for email body -small refactoring

+++ internal/app/service/user.go
@@ -1,8 +1,10 @@
 package service
 
 import (
-	"manga-explorer/internal/app/common/constant"
-	"manga-explorer/internal/app/common/status"
+	"fmt"
+	"log"
+	"manga-explorer/internal/common"
+	"manga-explorer/internal/common/status"
 	"manga-explorer/internal/domain/users"
 	"manga-explorer/internal/domain/users/dto"
 	"manga-explorer/internal/domain/users/mapper"
@@ -13,14 +15,17 @@ import (
 	"manga-explorer/internal/infrastructure/mail"
 	mailService "manga-explorer/internal/infrastructure/mail/service"
 	"manga-explorer/internal/util/containers"
+	"manga-explorer/internal/util/opt"
 	"time"
 )
 
-func NewUser(userRepo repository.IUser, verification service.IVerification, authentication service.IAuthentication, mail mailService.IMail) service.IUser {
-	return &userService{repo: userRepo, verifService: verification, mailService: mail, authService: authentication}
+func NewUser(config *common.Config, userRepo repository.IUser, verification service.IVerification, authentication service.IAuthentication, mail mailService.IMail, file fileService.IFile) service.IUser {
+	return &userService{config: config, repo: userRepo, verifService: verification, mailService: mail, fileService: file, authService: authentication}
 }
 
 type userService struct {
+	config *common.Config
+
 	repo         repository.IUser
 	verifService service.IVerification
 	authService  service.IAuthentication
@@ -32,7 +37,7 @@ func (u userService) UpdateProfileImage(input *dto.ProfileImageUpdateInput) stat
 	// Get user profile
 	profiles, err := u.repo.FindUserProfiles(input.UserId)
 	if err != nil {
-		return status.RepositoryError(err)
+		return status.RepositoryError(err, opt.New(status.PROFILE_NOT_FOUND))
 	}
 
 	// Upload new image
@@ -45,6 +50,7 @@ func (u userService) UpdateProfileImage(input *dto.ProfileImageUpdateInput) stat
 	if len(profiles.PhotoURL) > 0 {
 		stat = u.fileService.Delete(file.ProfileAsset, profiles.PhotoURL)
 		if stat.IsError() {
+			u.fileService.Delete(file.ProfileAsset, filename) // Delete uploaded image due to deletion error
 			return stat
 		}
 	}
@@ -52,70 +58,82 @@ func (u userService) UpdateProfileImage(input *dto.ProfileImageUpdateInput) stat
 	// Set metadata
 	profile := users.Profile{Id: profiles.Id, PhotoURL: filename, UpdatedAt: time.Now()}
 	err = u.repo.UpdateProfile(&profile)
-	return status.ConditionalRepository(err, status.UPDATED)
+	if err != nil {
+		u.fileService.Delete(file.ProfileAsset, filename)
+		return status.RepositoryError(err, opt.New(status.PROFILE_UPDATE_FAILED))
+	}
+	return status.Updated()
 }
 
 func (u userService) DeleteProfileImage(userId string) status.Object {
 	profiles, err := u.repo.FindUserProfiles(userId)
 	if err != nil {
-		return status.RepositoryError(err)
+		return status.RepositoryError(err, opt.New(status.PROFILE_NOT_FOUND))
 	}
+
+	// Do nothing when the profile has no image
+	if len(profiles.PhotoURL) == 0 {
+		return status.Deleted()
+	}
+
 	stat := u.fileService.Delete(file.ProfileAsset, profiles.PhotoURL)
 	if stat.IsError() {
 		return stat
 	}
 
-	profile := users.Profile{Id: profiles.Id, PhotoURL: "", UpdatedAt: time.Now()}
+	profile := users.Profile{Id: profiles.Id, PhotoURL: file.NoFile, UpdatedAt: time.Now()}
 	err = u.repo.UpdateProfile(&profile)
-	return status.ConditionalRepository(err, status.DELETED)
+	return status.ConditionalRepository(err, status.DELETED, opt.New(status.PROFILE_UPDATE_FAILED))
 }
 
 func (u userService) AddUser(input *dto.AddUserInput) status.Object {
 	userInput, err := mapper.MapAddUserInput(input)
 	if err != nil {
-		return status.InternalError()
+		return status.BadRequestError()
 	}
 	profileInput := mapper.MapAddProfileInput(&userInput, input)
 
 	err = u.repo.CreateUser(&userInput, &profileInput)
-	return status.ConditionalRepository(err, status.CREATED)
-}
-func (u userService) DeleteUser(userId string) status.Object {
-	err := u.repo.DeleteUser(userId)
-	return status.ConditionalRepository(err, status.DELETED)
-}
-func (u userService) GetAllUsers() ([]dto.UserResponse, status.Object) {
-	allUsers, err := u.repo.GetAllUsers()
-	if err != nil {
-		return nil, status.RepositoryError(err)
-	}
-	result := containers.CastSlicePtr(allUsers, mapper.ToUserResponse)
-	return result, status.Success()
+	return status.ConditionalRepositoryE(err, status.CREATED, opt.New(status.USER_CREATION_ALREADY_EXIST), opt.New(status.USER_CREATION_ALREADY_EXIST))
 }
 
-func (u userService) UpdateUser(input *dto.UpdateUserInput) status.Object {
+func (u userService) DeleteUser(userId string) status.Object {
+	err := u.repo.DeleteUser(userId)
+	return status.ConditionalRepository(err, status.DELETED, opt.New(status.USER_NOT_FOUND))
+}
+
+func (u userService) GetAllUsers() ([]dto.UserResponse, status.Object) {
+	allUsers, err := u.repo.GetAllUsers()
+	result := containers.CastSlicePtr(allUsers, mapper.ToUserResponse)
+	return result, status.ConditionalRepository(err, status.SUCCESS, opt.New(status.SUCCESS))
+}
+
+func (u userService) UpdateUser(input *dto.UserUpdateInput) status.Object {
 	usr := mapper.MapUserUpdateInput(input)
 	err := u.repo.UpdateUser(&usr)
-	return status.ConditionalRepository(err, status.UPDATED)
+	return status.ConditionalRepository(err, status.UPDATED, opt.New(status.USER_UPDATE_FAILED))
 }
-func (u userService) UpdateUserExtended(input *dto.UpdateUserExtendedInput) status.Object {
-	user, err := mapper.MapUpdateUserExtendedInput(input)
+
+func (u userService) UpdateUserExtended(input *dto.UserUpdateExtendedInput) status.Object {
+	user, err := mapper.MapUserUpdateExtendedInput(input)
 	if err != nil {
 		return status.InternalError()
 	}
 	err = u.repo.UpdateUser(&user)
-	return status.ConditionalRepository(err, status.UPDATED)
+	return status.ConditionalRepository(err, status.UPDATED, opt.New(status.USER_UPDATE_FAILED))
 }
-func (u userService) UpdateProfileExtended(input *dto.UpdateProfileExtendedInput) status.Object {
-	profile := mapper.MapUpdateProfileExtendedInput(input)
-	err := u.repo.UpdateProfile(&profile)
-	return status.ConditionalRepository(err, status.UPDATED)
+func (u userService) UpdateProfileExtended(input *dto.ProfileUpdateExtendedInput) status.Object {
+	profile := mapper.MapProfileUpdateExtendedInput(input)
+	err := u.repo.UpdateProfileByUserId(&profile)
+	return status.ConditionalRepository(err, status.UPDATED, opt.New(status.PROFILE_UPDATE_FAILED))
 }
+
 func (u userService) UpdateProfile(input *dto.ProfileUpdateInput) status.Object {
 	profile := mapper.MapProfileUpdateInput(input)
-	err := u.repo.UpdateProfile(&profile)
-	return status.ConditionalRepository(err, status.UPDATED)
+	err := u.repo.UpdateProfileByUserId(&profile)
+	return status.ConditionalRepository(err, status.UPDATED, opt.New(status.PROFILE_UPDATE_FAILED))
 }
+
 func (u userService) RegisterUser(input *dto.UserRegisterInput) (dto.UserResponse, status.Object) {
 	usr, err := mapper.MapUserRegisterInput(input)
 	if err != nil {
@@ -125,7 +143,7 @@ func (u userService) RegisterUser(input *dto.UserRegisterInput) (dto.UserRespons
 	profile := mapper.MapProfileRegisterInput(&usr, input)
 	err = u.repo.CreateUser(&usr, &profile)
 	if err != nil {
-		return dto.UserResponse{}, status.RepositoryError(err)
+		return dto.UserResponse{}, status.RepositoryError(err, opt.New(status.USER_CREATION_ALREADY_EXIST))
 	}
 
 	// Send email Verification
@@ -134,47 +152,51 @@ func (u userService) RegisterUser(input *dto.UserRegisterInput) (dto.UserRespons
 		return dto.UserResponse{}, stat
 	}
 
-	ml := &mail.Mail{
-		From:    constant.IssuerName,
-		To:      usr.Email,
-		Subject: "Email Verification",
-		Body:    verifResponse.Token,
+	m, err := mail.NewHTML("email-verification.gohtml",
+		fmt.Sprintf("%s/users/email-verif/%s", u.config.ApiDNS(1), verifResponse.Token))
+	if err != nil {
+		return dto.UserResponse{}, status.InternalError()
 	}
-	go u.mailService.SendEmail(ml)
+	m.Subject = "Email Verification"
+	m.Recipients = []string{usr.Email}
+	go func() {
+		err := u.mailService.SendEmail(m)
+		if err.IsError() {
+			log.Println("erer")
+		}
+	}()
 
-	return mapper.ToUserResponse(&usr), stat
+	return mapper.ToUserResponse(&usr), status.Created()
 }
+
 func (u userService) FindUserById(id string) (dto.UserResponse, status.Object) {
 	usr, err := u.repo.FindUserById(id)
 	if err != nil {
-		return dto.UserResponse{}, status.RepositoryError(err)
+		return dto.UserResponse{}, status.RepositoryError(err, opt.New(status.USER_NOT_FOUND))
 	}
 	return mapper.ToUserResponse(usr), status.Success()
 }
+
 func (u userService) FindUserByEmail(email string) (dto.UserResponse, status.Object) {
 	usr, err := u.repo.FindUserByEmail(email)
-	if err != nil {
-		return dto.UserResponse{}, status.RepositoryError(err)
-	}
-	return mapper.ToUserResponse(usr), status.Success()
+	return mapper.ToUserResponse(usr), status.RepositoryError(err, opt.New(status.USER_NOT_FOUND))
 }
+
 func (u userService) FindUserProfileById(userId string) (dto.ProfileResponse, status.Object) {
 	profile, err := u.repo.FindUserProfiles(userId)
-	if err != nil {
-		return dto.ProfileResponse{}, status.RepositoryError(err)
-	}
-	return mapper.ToProfileResponse(profile, u.fileService), status.Success()
+	return mapper.ToProfileResponse(profile, u.fileService), status.ConditionalRepository(err, status.SUCCESS, opt.New(status.PROFILE_NOT_FOUND))
 }
+
 func (u userService) ChangePassword(input *dto.ChangePasswordInput) status.Object {
 	// Get user
 	usr, err := u.repo.FindUserById(input.UserId)
 	if err != nil {
-		return status.RepositoryError(err)
+		return status.RepositoryError(err, opt.New(status.USER_NOT_FOUND))
 	}
 
-	// Check last password
+	// Match last password
 	if !usr.ValidatePassword(input.LastPassword) {
-		return status.Error(status.USER_LOGIN_ERROR)
+		return status.Error(status.USER_CHANGE_PASSWORD_WRONG_PASSWORD)
 	}
 
 	// Set new password
@@ -183,12 +205,13 @@ func (u userService) ChangePassword(input *dto.ChangePasswordInput) status.Objec
 		return status.InternalError()
 	}
 	err = u.repo.UpdateUser(&updateUser)
-	return status.ConditionalRepository(err, status.UPDATED)
+	return status.ConditionalRepository(err, status.UPDATED, opt.New(status.USER_UPDATE_FAILED))
 }
+
 func (u userService) RequestResetPassword(input *dto.ResetPasswordRequestInput) status.Object {
 	userResponse, err := u.repo.FindUserByEmail(input.Email)
 	if err != nil {
-		return status.RepositoryError(err)
+		return status.RepositoryError(err, opt.New(status.USER_NOT_FOUND))
 	}
 
 	verifResponse, stat := u.verifService.Request(userResponse.Id, users.UsageResetPassword)
@@ -196,42 +219,70 @@ func (u userService) RequestResetPassword(input *dto.ResetPasswordRequestInput) 
 		return stat
 	}
 
-	ml := &mail.Mail{
-		From:    constant.IssuerName,
-		To:      userResponse.Email,
-		Subject: "Reset Password",
-		Body:    verifResponse.Token,
+	m, err := mail.NewHTML("forgot-password.gohtml",
+		fmt.Sprintf("%s/users/reset-password/%s", u.config.DNS(), verifResponse.Token))
+	if err != nil {
+		return status.InternalError()
 	}
-	go u.mailService.SendEmail(ml)
+	m.Subject = "Reset Password"
+	m.Recipients = []string{userResponse.Email}
+	go u.mailService.SendEmail(m)
 	return status.Success()
 }
+
 func (u userService) ResetPassword(input *dto.ResetPasswordInput) status.Object {
-	// Find token
-	verifResponse, stat := u.verifService.Find(input.Token)
-	if stat.IsError() {
-		return stat
-	}
-
-	// Validate token
-	stat = u.verifService.Validate(&verifResponse)
-	if stat.IsError() {
-		return stat
-	}
-
-	// Immediately revoke verify token
-	stat = u.verifService.Remove(input.Token)
+	// Response token
+	resp, stat := u.verifService.Response(input.Token, users.UsageResetPassword)
 	if stat.IsError() {
 		return stat
 	}
 
 	// Logout all devices
-	u.authService.LogoutDevices(verifResponse.UserId)
+	stat = u.authService.LogoutDevices(resp.UserId)
+	if stat.IsError() {
+		return stat
+	}
 
 	// Set new password
-	updateUser, err := mapper.MapResetPasswordInput(input, verifResponse.UserId)
+	updateUser, err := mapper.MapResetPasswordInput(input, resp.UserId)
 	if err != nil {
 		return status.InternalError()
 	}
 	err = u.repo.UpdateUser(&updateUser)
-	return status.ConditionalRepository(err, status.SUCCESS)
+	return status.ConditionalRepository(err, status.SUCCESS, opt.New(status.USER_UPDATE_FAILED))
+}
+
+func (u userService) RequestEmailVerification(input *dto.VerifEmailRequestInput) status.Object {
+	userResponse, err := u.repo.FindUserById(input.UserId)
+	if err != nil {
+		return status.RepositoryError(err, opt.New(status.VERIFICATION_USER_NOT_EXISTS))
+	}
+
+	verifResponse, stat := u.verifService.Request(userResponse.Id, users.UsageVerifyEmail)
+	if stat.IsError() {
+		return stat
+	}
+
+	// Send email
+	m, err := mail.NewHTML("email-verification.gohtml",
+		fmt.Sprintf("%s/users/email-verif/%s", u.config.DNS(), verifResponse.Token))
+	if err != nil {
+		return status.InternalError()
+	}
+	m.Subject = "Email Verification"
+	m.Recipients = []string{userResponse.Email}
+	go u.mailService.SendEmail(m)
+
+	return status.Success()
+}
+
+func (u userService) VerifyEmail(input *dto.VerifyEmailInput) status.Object {
+	response, stat := u.verifService.Response(input.Token, users.UsageVerifyEmail)
+	if stat.IsError() {
+		return stat
+	}
+
+	updatedUser := mapper.MapVerifyEmailInput(response.UserId)
+	err := u.repo.UpdateUser(&updatedUser)
+	return status.ConditionalRepository(err, status.UPDATED, opt.New(status.USER_UPDATE_FAILED))
 }
